Bring back old JSON + remove default value
This commit sets the stage to start working on the new JSON
serializer, which will handle deserializing fields in an
unordered fashion.

Also, I deleted the "default value" parameter of the function
decode, since I believe it should be a method implemented
in the decoder itself.

I brought back the previous code we had for JSON (without
any dependences to json-AS), and modify it accordingly for
it to pass all tests. I also modified borsh, and added
back JSONBuffer (which is like JSON, but handling U8Arrs
instead of strings).