Moved defaultValue from function parameter to core.util.method
Until now, the deserializers would take a "defaultValue" as a
parameter in the decode_field method. Such defaultValue would
be the one used in case the `field` was not found in the
encoded structure. This was useful only in JSON, for which I
think it is better to remove it as a mandatory parameter, and
just use it if necesary.